Psuedo Auth implemented
Ready to add real auth xD
